Check PC-RP message lenght mathces data received.

+++ src/rp2040/network.c
@@ -15,7 +15,7 @@ int32_t get_UDP(
     uint8_t socket_num,
     uint16_t port,
     void* rx_buf,
-    uint8_t* data_received,
+    uint16_t* data_received,
     uint8_t* destip,
     uint16_t* destport)
 {
